Shows and episodes are integrated.
Now the site can load all shows first and when a user selects a show, they will be taken to all episodes of the show they picked.
I have also provided a navigation link for the user to be able to go back to all shows.
N.B: The app is still under construction.